+++ scripts/fade-unprepared.js
@@ -1,17 +1,16 @@
 export function setupLukasFadeUnprepared() {
 
-  Hooks.on("renderActorSheet", (app, html) => {
+  Hooks.on("renderBaseActorSheet", (app, html) => {
     // Rather than checking that it is a 5e sheet we check that it is a sheet with items and the correct classes
     if (!game.settings.get("dnd5e-scriptlets", "lukasFadeUnprepared")) return;
     
-    let items = html.find($(".items-list .item"));
+    let items = html.querySelectorAll(".items-list .item");
     for (let itemElement of items) {
       let id = itemElement.outerHTML.match(/data-item-id="(.*?)"/);
       if (!id) continue;
-      let item = app.object.items.get(id[1]);
-      if (!item) continue;
-
-      if (item.system?.preparation?.prepared === false && item.system?.preparation?.mode === "prepared" && item.system.level > 0) {
+      let item = app.document.items.get(id[1]);
+      if (!item || item.type !== "spell") continue;
+      if (item.system?.prepared === 0 && item.system?.method === "spell" && item.system.level > 0) {
         itemElement.classList.add("preparation-unprepared");
       }
     }