+++ scripts/module.js
@@ -1,9 +1,6 @@
 import { registerSettings } from "./settings.js";
 import { setupLukasItemRarityColors } from "./item-rarity-colors.js";
 import { setupLukasFadeUnprepared } from "./fade-unprepared.js";
-import { setupLegendaryRecharge } from "./legendaryRecharge.js";
-import { setupAutoItemRecharge } from "./autoItemRecharge.js";
-import { setupAutoRollUnlinkedHP } from "./autoRollUnlinkedHP.js";
 import { setupTokenResizer } from "./tokenResizer.js";
 import { setupAlternativeAdvantage } from "./alternativeAdvantage.js";
 import { setupCollapsibleActorSections } from "./collapsibleActorSections.js";
@@ -11,10 +8,8 @@ import { setupAmmoRecovery } from "./ammoRecovery.js";
 import { setupSilentInitiative } from "./silentInitiative.js";
 import { setupUpdateCreatedOrigins } from "./updateCreatedOrigins.js";
 import { setupSocket } from "./GMAction.js";
-import { setupGriddedGridless } from "./griddedGridless.js";
 import { initActorDispositionColors } from "./ActorDispositionColors.js";
 import { setupContainerHelpers } from "./containerHelper.js";
-import { setupEnableTokenDynamicRing } from "./enableTokenDynamicRing.js";
 import { setupTemplatePreview } from "./templatePreview.js";
 
 export let systemString;
@@ -48,7 +43,6 @@ Hooks.once("init", async function () {
 
 Hooks.once("setup", () => {
   registerSettings();
-
   console.log("dnd5e-scriptlets | doing setup");
 });
 
@@ -57,17 +51,12 @@ Hooks.once("ready", async function () {
   setupSocket();
   setupLukasItemRarityColors();
   setupLukasFadeUnprepared();
-  setupLegendaryRecharge();
-  setupAutoItemRecharge();
-  setupAutoRollUnlinkedHP();
   setupAlternativeAdvantage();
   setupCollapsibleActorSections();
   setupAmmoRecovery();
   setupSilentInitiative();
-  setupUpdateCreatedOrigins();
-  setupGriddedGridless();
+  // setupUpdateCreatedOrigins();
   setupContainerHelpers();
-  setupEnableTokenDynamicRing();
   const module = game.modules.get("dnd5e-scriptlets");
   if (module) {
     module.api = globalThis.dnd5eScriptlets.api;